backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
numpy>=1.26.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
tldextract>=5.0.0
# Database
sqlalchemy>=2.0.0
# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
# PostgreSQL support (for production)
psycopg2-binary>=2.9.0
